fields are now strings + first iteration of caller_id search tested
every field in DB struct is now a string instead of an array of char to help me in the future with memory bounds. Now I don't have to know the array's size.

Tested the search_exact_caller_id() function and seemed to behave properly in both positive and negative scenarios